+++ bijection-core/src/main/scala/com/twitter/bijection/Conversion.scala
@@ -17,6 +17,8 @@ limitations under the License.
 package com.twitter.bijection
 
 import java.io.Serializable
+import scala.util.Success
+
 /**
  * Convert allows the user to convert an instance of type A to type B given an implicit Conversion
  * that goes between the two.
@@ -43,8 +45,15 @@ trait Conversion[A, B] extends Serializable {
 
 trait CrazyLowPriorityConversion extends Serializable {
   // If you want to an Option[B]
-  implicit def fromInjectionInverse[A,B](implicit inj: Injection[B,A]): Conversion[A,Option[B]] =
+  implicit def fromInjectionOptInverse[A,B](implicit inj: Injection[B,A]): Conversion[A, Option[B]] =
   new Conversion[A,Option[B]] {
+    def apply(a: A) = inj.invert(a) match {
+      case Success(value) => Some(value)
+      case _ => None
+    }
+  }
+  implicit def fromInjectionInverse[A,B](implicit inj: Injection[B,A]): Conversion[A,Attempt[B]] =
+  new Conversion[A,Attempt[B]] {
     def apply(a: A) = inj.invert(a)
   }
   implicit def fromBijectionInv[A,B](implicit fn: ImplicitBijection[B,A]) = new Conversion[A,B] {