+++ bijection-core/src/test/scala/com/twitter/bijection/BaseProperties.scala
@@ -20,6 +20,7 @@ import org.scalacheck.{ Arbitrary, Properties }
 import org.scalacheck.Prop.forAll
 
 import scala.math.Equiv
+import scala.util.Success
 import java.util.Arrays
 
 trait BaseProperties {
@@ -42,20 +43,20 @@ trait BaseProperties {
         b != null &&
         { val bofa = inj.invert(b)
           bofa != null &&
-          bofa.isDefined &&
+          bofa.isSuccess &&
           eqa.equiv(bofa.get, a)
         }
     }
 
  def invertIsStrict[A,B](implicit arbb: Arbitrary[B], inj: Injection[A,B], eqb: Equiv[B]) =
    forAll { (b: B) =>
-     inj.invert(b)
-       .map { aofb =>
+     inj.invert(b) match {
+       case Success(aofb) =>
          assert(aofb != null, "aofb was null")
          eqb.equiv(b, inj(aofb))
-       }
-       .getOrElse(true)
+       case _ => true // the failing case of None previously returned true?
      }
+   }
 
   def isInjection[A,B](implicit a: Arbitrary[A],
     inj: Injection[A, B], barb: Arbitrary[B], eqa: Equiv[A], eqb: Equiv[B]) =