+++ project/Build.scala
@@ -5,12 +5,15 @@ import Keys._
 import sbtgitflow.ReleasePlugin._
 import com.typesafe.tools.mima.plugin.MimaPlugin.mimaDefaultSettings
 import com.typesafe.tools.mima.plugin.MimaKeys.previousArtifact
+import com.typesafe.sbt.osgi.SbtOsgi._
 
 object BijectionBuild extends Build {
-  val sharedSettings = Project.defaultSettings ++ releaseSettings ++ Seq(
+  val sharedSettings = Project.defaultSettings ++ releaseSettings ++ osgiSettings ++ Seq(
     organization := "com.twitter",
 
-    crossScalaVersions := Seq("2.9.2", "2.10.0"),
+    crossScalaVersions := Seq("2.9.3", "2.10.2"),
+
+    scalaVersion := "2.9.3",
 
     javacOptions ++= Seq("-source", "1.6", "-target", "1.6"),
 
@@ -29,6 +32,12 @@ object BijectionBuild extends Build {
     parallelExecution in Test := true,
 
     scalacOptions ++= Seq("-unchecked", "-deprecation"),
+
+    OsgiKeys.importPackage <<= scalaVersion { sv => Seq("""scala.*;version="$<range;[==,=+);%s>"""".format(sv)) },
+
+    OsgiKeys.importPackage ++= Seq("com.twitter.bijection.*;version=\"[${Bundle-Version}, ${Bundle-Version}]\"", "*"),
+
+    OsgiKeys.additionalHeaders := Map("-removeheaders" -> "Include-Resource,Private-Package"),
 
     // Publishing options:
     publishMavenStyle := true,
@@ -81,8 +90,10 @@ object BijectionBuild extends Build {
   // This returns the youngest jar we released that is compatible with the current
   def youngestForwardCompatible(subProj: String) = {
     if(subProj == "netty") None // This is new. Update after next version
-    else Some("com.twitter" % ("bijection-" + subProj + "_2.9.2") % "0.3.0")
+    else Some("com.twitter" % ("bijection-" + subProj + "_2.9.2") % "0.4.0")
   }
+
+  def osgiExportAll(packs: String*) = OsgiKeys.exportPackage := packs.map(_ + ".*;version=${Bundle-Version}")
 
   lazy val bijection = Project(
     id = "bijection",
@@ -103,6 +114,7 @@ object BijectionBuild extends Build {
               bijectionClojure,
               bijectionNetty)
 
+  /** No dependencies in bijection other than java + scala */
   lazy val bijectionCore = Project(
     id = "bijection-core",
     base = file("bijection-core"),
@@ -110,8 +122,8 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-core",
     previousArtifact := youngestForwardCompatible("core"),
+    osgiExportAll("com.twitter.bijection"),
     libraryDependencies ++= Seq(
-        "commons-codec" % "commons-codec" % "1.7",
         "com.novocode" % "junit-interface" % "0.10-M1" % "test",
         "org.scalatest" %% "scalatest" % "1.9.1" % "test"
     )
@@ -124,6 +136,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-protobuf",
     previousArtifact := youngestForwardCompatible("protobuf"),
+    osgiExportAll("com.twitter.bijection.protobuf"),
     libraryDependencies += "com.google.protobuf" % "protobuf-java" % "2.4.1"
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 
@@ -136,6 +149,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-thrift",
     previousArtifact := youngestForwardCompatible("thrift"),
+    osgiExportAll("com.twitter.bijection.thrift"),
     libraryDependencies ++= Seq(
       "org.apache.thrift" % "libthrift" % "0.6.1" exclude("junit", "junit"),
       jsonParser
@@ -149,6 +163,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-guava",
     previousArtifact := youngestForwardCompatible("guava"),
+    osgiExportAll("com.twitter.bijection.guava"),
     libraryDependencies ++= Seq(
       // This dependency is required due to a bug with guava 13.0, detailed here:
       // http://code.google.com/p/guava-libraries/issues/detail?id=1095
@@ -164,6 +179,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-scrooge",
     previousArtifact := youngestForwardCompatible("scrooge"),
+    osgiExportAll("com.twitter.bijection.scrooge"),
     libraryDependencies ++= Seq(
       "org.apache.thrift" % "libthrift" % "0.6.1" exclude("junit", "junit"),
       "com.twitter" % "scrooge-runtime" % "3.0.4"
@@ -177,6 +193,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-json",
     previousArtifact := youngestForwardCompatible("json"),
+    osgiExportAll("com.twitter.bijection.json"),
     libraryDependencies += jsonParser
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 
@@ -187,7 +204,12 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-algebird",
     previousArtifact := youngestForwardCompatible("algebird"),
-    libraryDependencies += "com.twitter" %% "algebird-core" % "0.1.9"
+    osgiExportAll("com.twitter.bijection.algebird"),
+    libraryDependencies += "com.twitter" %% "algebird-core" % "0.1.9" cross CrossVersion.binaryMapped {
+      case "2.9.3" => "2.9.2" // TODO: hack because twitter hasn't built things agaisnt 2.9.3
+      case version if version startsWith "2.10" => "2.10" // TODO: hack because sbt is broken
+      case x       => x
+    }
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 
   lazy val bijectionUtil = Project(
@@ -197,7 +219,12 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-util",
     previousArtifact := youngestForwardCompatible("util"),
-    libraryDependencies += "com.twitter" %% "util-core" % "6.2.0"
+    osgiExportAll("com.twitter.bijection.util"),
+    libraryDependencies += "com.twitter" %% "util-core" % "6.2.0" cross CrossVersion.binaryMapped {
+      case "2.9.3" => "2.9.2" // TODO: hack because twitter hasn't built things agaisnt 2.9.3
+      case version if version startsWith "2.10" => "2.10" // TODO: hack because sbt is broken
+      case x       => x
+    }
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 
   lazy val bijectionClojure = Project(
@@ -207,6 +234,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-clojure",
     previousArtifact := youngestForwardCompatible("clojure"),
+    osgiExportAll("com.twitter.bijection.clojure"),
     libraryDependencies += "org.clojure" % "clojure" % "1.4.0"
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 
@@ -217,6 +245,7 @@ object BijectionBuild extends Build {
   ).settings(
     name := "bijection-netty",
     previousArtifact := youngestForwardCompatible("netty"),
+    osgiExportAll("com.twitter.bijection.netty"),
     libraryDependencies += "io.netty" % "netty" % "3.5.5.Final"
   ).dependsOn(bijectionCore % "test->test;compile->compile")
 }